+++ rsvim_core/src/js/module/module_map.rs
@@ -1,0 +1,125 @@
+//! Module map.
+
+use crate::js::module::es_module::*;
+use crate::js::module::{ModulePath, ModuleStatus};
+use crate::prelude::*;
+
+use std::collections::LinkedList;
+
+#[derive(Debug, Clone)]
+/// Import kind.
+pub enum ImportKind {
+  // Loading static imports.
+  Static,
+  // Loading a dynamic import.
+  Dynamic(v8::Global<v8::PromiseResolver>),
+}
+
+#[derive(Debug)]
+/// Module graph.
+pub struct ModuleGraph {
+  pub kind: ImportKind,
+  pub root_rc: EsModuleRc,
+  pub same_origin: LinkedList<v8::Global<v8::PromiseResolver>>,
+}
+
+rc_refcell_ptr!(ModuleGraph);
+
+impl ModuleGraph {
+  // Initializes a new graph resolving a static import.
+  pub fn static_import(path: &str) -> ModuleGraph {
+    // Create an ES module instance.
+    let module = EsModule::to_rc(EsModule::new(
+      path.into(),
+      ModuleStatus::Fetching,
+      vec![],
+      None,
+      false,
+    ));
+
+    Self {
+      kind: ImportKind::Static,
+      root_rc: module,
+      same_origin: LinkedList::new(),
+    }
+  }
+
+  // Initializes a new graph resolving a dynamic import.
+  pub fn dynamic_import(path: &str, promise: v8::Global<v8::PromiseResolver>) -> ModuleGraph {
+    // Create an ES module instance.
+    let module = EsModule::to_rc(EsModule::new(
+      path.into(),
+      ModuleStatus::Fetching,
+      vec![],
+      None,
+      true,
+    ));
+
+    Self {
+      kind: ImportKind::Dynamic(promise),
+      root_rc: module,
+      same_origin: LinkedList::new(),
+    }
+  }
+}
+
+/// Module map.
+/// It maintains all the modules inside js runtime, including already resolved and pending
+/// fetching.
+pub struct ModuleMap {
+  pub main: Option<ModulePath>,
+  pub index: HashMap<ModulePath, v8::Global<v8::Module>>,
+  pub seen: HashMap<ModulePath, ModuleStatus>,
+  pub pending: Vec<ModuleGraphRc>,
+}
+
+impl ModuleMap {
+  // Creates a new module-map instance.
+  pub fn new() -> ModuleMap {
+    Self {
+      main: None,
+      index: HashMap::new(),
+      seen: HashMap::new(),
+      pending: vec![],
+    }
+  }
+
+  // Inserts a compiled ES module to the map.
+  pub fn insert(&mut self, path: &str, module: v8::Global<v8::Module>) {
+    // No main module has been set, so let's update the value.
+    if self.main.is_none() && std::fs::metadata(path).is_ok() {
+      self.main = Some(path.into());
+    }
+    self.index.insert(path.into(), module);
+  }
+
+  // Returns if there are still pending imports to be loaded.
+  pub fn has_pending_imports(&self) -> bool {
+    !self.pending.is_empty()
+  }
+
+  // Returns a v8 module reference from me module-map.
+  pub fn get(&self, key: &str) -> Option<v8::Global<v8::Module>> {
+    self.index.get(key).cloned()
+  }
+
+  // Returns a specifier by a v8 module.
+  pub fn get_path(&self, module: v8::Global<v8::Module>) -> Option<ModulePath> {
+    self
+      .index
+      .iter()
+      .find(|(_, m)| **m == module)
+      .map(|(p, _)| p.clone())
+  }
+
+  // Returns the main entry point.
+  pub fn main(&self) -> Option<ModulePath> {
+    self.main.clone()
+  }
+}
+
+impl Default for ModuleMap {
+  fn default() -> Self {
+    ModuleMap::new()
+  }
+}
